Fixed a warning with unsigned to signed integer comparison
git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/Phosphoros/trunk@6544 463500fe-d6be-4184-9b20-47717e75ecac

+++ PhzDataModel/PhzDataModel/serialization/PhotometryGrid.h
@@ -70,7 +70,7 @@ void load(Archive& ar, Euclid::PhzDataModel::PhotometryGrid& grid, const unsigne
   auto filter_names_ptr = std::make_shared<std::vector<std::string>>(std::move(filter_names));
   for (auto& cell : grid) {
     std::vector<Euclid::SourceCatalog::FluxErrorPair> phot_values;
-    for (int i=0; i< filter_names_ptr->size(); ++i) {
+    for (size_t i=0; i< filter_names_ptr->size(); ++i) {
       double flux;
       double error;
       ar >> flux >> error;
